+++ NEWS.md
@@ -3,6 +3,8 @@
 * Braintree::Transaction.void updates the existing sale transaction instead of
   creating a new transaction of type Braintree::Transaction::Status::Voided
 
+* Preserve `redirect_url` query parameters for transparent redirect (#36)
+
 # 0.2.1
 * Fake refunds via `Braintree::Transaction.refund` and
   `Braintree::CreditCard.refund`.